+++ src/editor/gui/view_map_stamp.cpp
@@ -19,6 +19,7 @@
 #include <editor/gui/view_map_stamp.h>
 #include <editor/gui/image_picker.h>
 #include <editor/gui/tilingmaterial_picker.h>
+#include <editor/gui/zone_picker.h>
 #include <editor/components/gui.h>
 #include <client/components/assetsrenderer.h>
 #include <client/gui/popup.h>
@@ -152,254 +153,41 @@ public:
 class CPopup_ZonePalette : public gui::CPopup
 {
 protected:
-	class CZoneIcon : public gui::CWidget
-	{
-	protected:
-		CGuiEditor* m_pAssetsEditor;
-		CAssetPath m_ZoneTypePath;
-		CSubPath m_IndexSubPath;
-		array2d<CAsset_MapZoneTiles::CTile> m_Tiles;
-		int m_Number;
-		
-	public:
-		CZoneIcon(CGuiEditor *pAssetsEditor, CAssetPath ZoneTypePath, CSubPath IndexSubPath) :
-			gui::CWidget(pAssetsEditor),
-			m_pAssetsEditor(pAssetsEditor),
-			m_ZoneTypePath(ZoneTypePath),
-			m_IndexSubPath(IndexSubPath)
-		{
-			int Index = 0;
-			if(m_IndexSubPath.IsNotNull())
-			{
-				const CAsset_ZoneType* pZoneType = AssetsManager()->GetAsset<CAsset_ZoneType>(m_ZoneTypePath);
-				if(pZoneType && pZoneType->IsValidIndex(m_IndexSubPath))
-					Index = m_IndexSubPath.GetId();
-			}
-			
-			m_Tiles.resize(5, 5);
-			
-			m_Tiles.get_clamp(0, 0).SetIndex(0);
-			m_Tiles.get_clamp(1, 0).SetIndex(0);
-			m_Tiles.get_clamp(2, 0).SetIndex(0);
-			m_Tiles.get_clamp(3, 0).SetIndex(0);
-			m_Tiles.get_clamp(4, 0).SetIndex(0);
-			
-			m_Tiles.get_clamp(0, 1).SetIndex(0);
-			m_Tiles.get_clamp(1, 1).SetIndex(Index);
-			m_Tiles.get_clamp(2, 1).SetIndex(Index);
-			m_Tiles.get_clamp(3, 1).SetIndex(0);
-			m_Tiles.get_clamp(4, 1).SetIndex(0);
-			
-			m_Tiles.get_clamp(0, 2).SetIndex(0);
-			m_Tiles.get_clamp(1, 2).SetIndex(Index);
-			m_Tiles.get_clamp(2, 2).SetIndex(Index);
-			m_Tiles.get_clamp(3, 2).SetIndex(Index);
-			m_Tiles.get_clamp(4, 2).SetIndex(0);
-			
-			m_Tiles.get_clamp(0, 3).SetIndex(0);
-			m_Tiles.get_clamp(1, 3).SetIndex(Index);
-			m_Tiles.get_clamp(2, 3).SetIndex(Index);
-			m_Tiles.get_clamp(3, 3).SetIndex(Index);
-			m_Tiles.get_clamp(4, 3).SetIndex(0);
-			
-			m_Tiles.get_clamp(0, 4).SetIndex(0);
-			m_Tiles.get_clamp(1, 4).SetIndex(0);
-			m_Tiles.get_clamp(2, 4).SetIndex(0);
-			m_Tiles.get_clamp(3, 4).SetIndex(0);
-			m_Tiles.get_clamp(4, 4).SetIndex(0);
-		}
-		
-		virtual void UpdateBoundingSize()
-		{
-			m_BoundingSizeRect.BSMinimum(32.0f*2, 32.0f*2);
-		}
-		
-		virtual void Render()
-		{
-			Graphics()->ClipPush(m_DrawRect.x, m_DrawRect.y, m_DrawRect.w, m_DrawRect.h);
-			
-			CMapRenderer MapRenderer(ClientKernel());
-			
-			MapRenderer.SetTime(0.0f);
-			MapRenderer.SetLocalTime(time_get()/(float)time_freq());
-			MapRenderer.SetCanvas(m_DrawRect, vec2(m_DrawRect.x + m_DrawRect.w/2, m_DrawRect.y + m_DrawRect.h/2));
-			MapRenderer.SetCamera(0.0f, 1.0f);
-			
-			MapRenderer.RenderTiles_Zone(m_ZoneTypePath, m_Tiles, vec2(-32.0f*2.5f, -32.0f*2.5f), 1.0f, true);
-			
-			Graphics()->ClipPop();
-		}
-	};
-	
-	class CZoneButton : public gui::CButton
+	class CPaletteZonePicker : public CZonePicker
 	{
 	protected:
 		CPopup_ZonePalette* m_pPopup;
 		CCursorTool_MapStamp* m_pCursorTool;
-		CAssetPath m_ZoneTypePath;
-		CSubPath m_IndexSubPath;
 		
-	public:
-		CZoneButton(CGuiEditor *pAssetsEditor, CCursorTool_MapStamp* pCursorTool, CPopup_ZonePalette* pPopup, CAssetPath ZoneTypePath, CSubPath IndexSubPath) :
-			gui::CButton(pAssetsEditor, ""),
-			m_pPopup(pPopup),
-			m_pCursorTool(pCursorTool),
-			m_ZoneTypePath(ZoneTypePath),
-			m_IndexSubPath(IndexSubPath)
+	protected:
+		virtual void OnZonePicked(CSubPath IndexPath)
 		{
-			//~ SetText("Empty");
-			
-			if(IndexSubPath.IsNotNull())
-			{
-				//~ const CAsset_ZoneType* pZoneType = AssetsManager()->GetAsset<CAsset_ZoneType>(ZoneTypePath);
-				//~ if(pZoneType && pZoneType->IsValidIndex(m_IndexSubPath) && pZoneType->GetIndexUsed(m_IndexSubPath))
-					//~ SetText(pZoneType->GetIndexTitle(m_IndexSubPath));
-			}
-				
-			
-			SetButtonStyle(pAssetsEditor->m_Path_Button_PaletteIcon);
-			SetIconWidget(new CZoneIcon(pAssetsEditor, ZoneTypePath, m_IndexSubPath));
-		}
-		
-		virtual void MouseClickAction()
-		{
-			m_pCursorTool->PaletteCallback_SelectZoneType(m_ZoneTypePath, m_IndexSubPath, m_pPopup->m_DataInt);
+			m_pCursorTool->PaletteCallback_SelectZoneType(m_ZoneTypePath, IndexPath, m_DataInt);
 			m_pPopup->Close();
 		}
-		
-		virtual void OnMouseMove()
+	
+	public:
+		CPaletteZonePicker(CCursorTool_MapStamp* pCursorTool, CPopup_ZonePalette* pPopup, CAssetPath ZoneTypePath) :
+			CZonePicker(pCursorTool->AssetsEditor(), ZoneTypePath, pCursorTool->ViewMap()->GetMemorizedZoneData(ZoneTypePath)),
+			m_pPopup(pPopup),
+			m_pCursorTool(pCursorTool)
 		{
-			if(m_pPopup->m_pZoneTitle && m_pPopup->m_pZoneDesc && m_VisibilityRect.IsInside(Context()->GetMousePos()))
-			{
-				const CAsset_ZoneType* pZoneType = AssetsManager()->GetAsset<CAsset_ZoneType>(m_ZoneTypePath);
-				if(pZoneType && pZoneType->IsValidIndex(m_IndexSubPath) && pZoneType->GetIndexUsed(m_IndexSubPath))
-				{
-					m_pPopup->m_pZoneTitle->SetText(pZoneType->GetIndexTitle(m_IndexSubPath));
-					m_pPopup->m_pZoneDesc->SetText(pZoneType->GetIndexDescription(m_IndexSubPath));
-				}
-				else
-				{
-					m_pPopup->m_pZoneTitle->SetText(_LSTRING("Empty"));
-					m_pPopup->m_pZoneDesc->SetText(_LSTRING("Can be used as an eraser"));
-				}
-			}
 			
-			gui::CButton::OnMouseMove();
 		}
 	};
 	
 protected:
-	CAssetPath m_ZoneTypePath;
-	std::vector<int> m_DataInt;
-	gui::CLabel* m_pZoneTitle;
-	gui::CLabel* m_pZoneDesc;
+	CGuiEditor* m_pAssetsEditor;
 	
 public:
 	CPopup_ZonePalette(CGuiEditor* pAssetsEditor, CCursorTool_MapStamp* pCursorTool, const gui::CRect& CreatorRect, CAssetPath ZoneTypePath) :
 		gui::CPopup(pAssetsEditor, CreatorRect, CreatorRect.w-16, CreatorRect.h-16, gui::CPopup::ALIGNMENT_INNER),
-		m_ZoneTypePath(ZoneTypePath),
-		m_pZoneTitle(NULL),
-		m_pZoneDesc(NULL)
+		m_pAssetsEditor(pAssetsEditor)
 	{
-		SetBoxStyle(pAssetsEditor->m_Path_Box_DialogPanel);
+		SetBoxStyle(m_pAssetsEditor->m_Path_Box_Dialog);
 		
-		const CAsset_ZoneType* pZoneType = AssetsManager()->GetAsset<CAsset_ZoneType>(m_ZoneTypePath);
-		if(!pZoneType)
-			return;
-		
-		gui::CVListLayout* pVList = new gui::CVListLayout(Context());
-		Add(pVList);
-		
-		//Description of the current zone
-		{
-			gui::CVListLayout* pHeader = new gui::CVListLayout(Context());
-			pHeader->SetBoxStyle(pAssetsEditor->m_Path_Box_PanelHeader);
-			pVList->Add(pHeader);
-			
-			m_pZoneTitle = new gui::CLabel(Context(), _LSTRING("Select a zone type by clicking on it"));
-			m_pZoneTitle->SetLabelStyle(pAssetsEditor->m_Path_Label_Group);
-			pHeader->Add(m_pZoneTitle, false);
-			m_pZoneDesc = new gui::CLabel(Context(), _LSTRING(""));
-			pHeader->Add(m_pZoneDesc, false);
-		}
-		
-		gui::CHPanelLayout* pPanels = new gui::CHPanelLayout(Context());
-		pVList->Add(pPanels, true);
-		
-		//List of zones
-		{
-			gui::CBoxLayout* pList = new gui::CBoxLayout(Context());
-			pPanels->Add(pList, -1);
-			pList->SetBoxStyle(pAssetsEditor->m_Path_Box_Panel);
-			
-			if(!pZoneType->IsValidIndex(CAsset_ZoneType::SubPath_Index(0)) || !pZoneType->GetIndexUsed(CAsset_ZoneType::SubPath_Index(0)))
-				pList->Add(new CZoneButton(pAssetsEditor, pCursorTool, this, m_ZoneTypePath, CSubPath::Null()));
-			
-			CAsset_ZoneType::CIteratorGroup GroupIter;
-			CAsset_ZoneType::CIteratorIndex IndexIter;
-			
-			//Show default group
-			bool UnknownGroupNeed = false;
-			for(IndexIter = pZoneType->BeginIndex(); IndexIter != pZoneType->EndIndex(); ++IndexIter)
-			{
-				if(pZoneType->GetIndexUsed(*IndexIter))
-				{
-					if(pZoneType->GetIndexGroup(*IndexIter) == -1)
-						pList->Add(new CZoneButton(pAssetsEditor, pCursorTool, this, m_ZoneTypePath, *IndexIter));
-					else if(pZoneType->IsValidGroup(CAsset_ZoneType::SubPath_Group(pZoneType->GetIndexGroup(*IndexIter))))
-						UnknownGroupNeed = true;
-				}
-			}
-			
-			//Show groups
-			for(GroupIter = pZoneType->BeginGroup(); GroupIter != pZoneType->EndGroup(); ++GroupIter)
-			{
-				gui::CLabel* pLabel = new gui::CLabel(Context(), pZoneType->GetGroup(*GroupIter));
-				pLabel->SetLabelStyle(pAssetsEditor->m_Path_Label_Group);
-				pList->Add(pLabel, true);
-				
-				for(IndexIter = pZoneType->BeginIndex(); IndexIter != pZoneType->EndIndex(); ++IndexIter)
-				{
-					if(pZoneType->GetIndexUsed(*IndexIter) && pZoneType->GetIndexGroup(*IndexIter) == (*GroupIter).GetId())
-						pList->Add(new CZoneButton(pAssetsEditor, pCursorTool, this, m_ZoneTypePath, *IndexIter));
-				}
-			}
-			
-			//Show unknown groups
-			if(UnknownGroupNeed)
-			{
-				gui::CLabel* pLabel = new gui::CLabel(Context(), _LSTRING("Other zones"));
-				pLabel->SetLabelStyle(pAssetsEditor->m_Path_Label_Group);
-				pList->Add(pLabel, true);
-				
-				for(IndexIter = pZoneType->BeginIndex(); IndexIter != pZoneType->EndIndex(); ++IndexIter)
-				{
-					if(pZoneType->GetIndexUsed(*IndexIter) && pZoneType->GetIndexGroup(*IndexIter) != -1 && !pZoneType->IsValidGroup(CAsset_ZoneType::SubPath_Group(pZoneType->GetIndexGroup(*IndexIter))))
-						pList->Add(new CZoneButton(pAssetsEditor, pCursorTool, this, m_ZoneTypePath, *IndexIter));
-				}
-			}
-		}
-		
-		//DataInt editor
-		if(pZoneType->GetDataIntArraySize())
-		{
-			m_DataInt.resize(pZoneType->GetDataIntArraySize());
-			
-			gui::CVScrollLayout* pDataList = new gui::CVScrollLayout(Context());
-			pPanels->Add(pDataList, 200);
-			pDataList->SetBoxStyle(pAssetsEditor->m_Path_Box_Panel);
-			
-			CAsset_ZoneType::CIteratorDataInt Iter;
-			for(Iter = pZoneType->BeginDataInt(); Iter != pZoneType->EndDataInt(); ++Iter)
-			{
-				m_DataInt[(*Iter).GetId()] = pZoneType->GetDataIntDefaultValue(*Iter);
-				
-				gui::CHListLayout* pHList = new gui::CHListLayout(Context());
-				pHList->Add(new gui::CLabel(Context(), pZoneType->GetDataIntTitle(*Iter)), true);
-				pHList->Add(new gui::CBoundedExternalIntegerEdit(Context(), &m_DataInt[(*Iter).GetId()], pZoneType->GetDataIntMinValue(*Iter), pZoneType->GetDataIntMaxValue(*Iter)), true);
-				pDataList->Add(pHList, false);
-			}
-		}
+		CPaletteZonePicker* pPicker = new CPaletteZonePicker(pCursorTool, this, ZoneTypePath);
+		Add(pPicker);
 	}
 	
 	virtual void OnButtonClick(int Button)
@@ -1847,6 +1635,8 @@ void CCursorTool_MapStamp::PaletteCallback_SelectZoneType(CAssetPath ZoneTypePat
 		m_IntDataSelection.resize(1, 1, pZoneType->GetDataIntArraySize());
 		for(unsigned int d=0; d<DataInt.size(); d++)
 			m_IntDataSelection.set_clamp(1, 1, d, DataInt[d]);
+		
+		ViewMap()->MemorizeZoneData(ZoneTypePath, DataInt);
 	}
 		
 	m_SelectionEnabled = true;